requirements.txt
asgiref==3.9.2
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
